Move EntriesCursor tests to unit.rs
